Remove Score struct wrapper, rename Traversal to Score
